requirements.txt
aiogram==3.2.0
aiohttp==3.9.1
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3